fix: consolidate logical-move cron service test for pkg-file-count
Move workerless cron logical-move coverage into poller_watcher_test.go so
pkg/service stays within the 15-file package limit after story 004 tests.
